[refactor-1021]: вынес функцию загрузки из компонента
