+++ frontend/src/widgets/cover-letter/index.ts
@@ -1,5 +1,4 @@
 export { CoverLetterEditor } from './ui/CoverLetterEditor'
 export { CoverLetterPreview } from './ui/CoverLetterPreview'
-export { downloadCoverLetterJson } from './lib/downloadCoverLetterJson'
-export { CoverLetterProvider } from './lib/CoverLetterProvider'
+export { CoverLetterProvider } from './context/CoverLetterProvider'
 export { CoverLetterExportButton } from './ui/CoverLetterExportButton'
